Refactor some math OCR code to the ocr.py provider
