+++ tests/provider/test_ocr.py
@@ -2,6 +2,7 @@ import unittest
 from mock import patch
 from provider import ocr
 from tests import settings_mock
+from tests.activity import test_activity_data
 from tests.activity.classes_mock import FakeLogger, FakeResponse
 
 
@@ -104,3 +105,111 @@ class TestMathpixPostRequest(unittest.TestCase):
             verify_ssl,
             self.logger,
         )
+
+
+class TestOcrFiles(unittest.TestCase):
+    "test ocr_files()"
+
+    def setUp(self):
+        self.caller_name = "AcceptedSubmissionPeerReviewOcr"
+        self.logger = FakeLogger()
+        self.file_name = "sa1-inf1.jpg"
+        self.file_path = "tests/files_source/digests/outbox/99999/digest-99999.jpg"
+        self.file_to_path_map = {self.file_name: self.file_path}
+        self.identifier = "test.zip"
+
+    @patch("provider.ocr.requests.post")
+    def test_ocr_files(self, fake_request):
+        "test a request to the ocr endpoint but mocking the requests.post"
+        options_type = "math"
+        response_json = {}
+        response_status = 200
+        response = FakeResponse(response_status)
+        response.response_json = response_json
+        fake_request.return_value = response
+        expected_result = {self.file_name: response_json}
+        # invoke
+        result = ocr.ocr_files(
+            self.file_to_path_map,
+            options_type,
+            settings_mock,
+            self.caller_name,
+            self.logger,
+            self.identifier,
+        )
+        # assert
+        self.assertDictEqual(result, expected_result)
+
+    @patch("requests.post")
+    def test_table_ocr_files(self, fake_request):
+        "test the options_type table"
+        options_type = "table"
+        response_json = {"data": [{"type": "tsv", "value": ""}]}
+        response_status = 200
+        response = FakeResponse(response_status, response_json)
+        # response.response_json = response_json
+        fake_request.return_value = response
+        expected_result = {self.file_name: response_json}
+        # invoke
+        result = ocr.ocr_files(
+            self.file_to_path_map,
+            options_type,
+            settings_mock,
+            self.caller_name,
+            self.logger,
+            self.identifier,
+        )
+        # assert
+        self.assertDictEqual(result, expected_result)
+
+    @patch("requests.post")
+    def test_failure_status_code(self, fake_request):
+        "test exception catching of a non-success status code response"
+        options_type = "math"
+        response_json = {}
+        response_status = 500
+        response = FakeResponse(response_status)
+        response.response_json = response_json
+        fake_request.return_value = response
+        expected_result = {}
+        excepted_log_message = (
+            "%s, exception posting to Mathpix API endpoint, file_name %s: "
+            "Error in mathpix_post_request %s to Mathpix API: %s\nNone"
+            % (self.caller_name, self.file_name, self.file_path, response_status)
+        )
+        # invoke
+        result = ocr.ocr_files(
+            self.file_to_path_map,
+            options_type,
+            settings_mock,
+            self.caller_name,
+            self.logger,
+            self.identifier,
+        )
+        # assert
+        self.assertDictEqual(result, expected_result)
+        # test for logging
+        self.assertEqual(self.logger.logexception, excepted_log_message)
+
+
+EQUATION_DATA = test_activity_data.EXAMPLE_OCR_RESPONSE_JSON.get("data")
+
+
+class TestMathDataParts(unittest.TestCase):
+    "test math_data_parts()"
+
+    def test_math_data_parts(self):
+        "test expected input"
+        math_data = EQUATION_DATA
+        mathml_data, latex_data = ocr.math_data_parts(math_data)
+        self.assertTrue(isinstance(mathml_data, dict))
+        self.assertTrue(isinstance(latex_data, dict))
+        self.assertEqual(list(mathml_data.keys()), ["type", "value"])
+        self.assertEqual(list(latex_data.keys()), ["type", "value"])
+
+    def test_empty_list(self):
+        "test when data is an empty list"
+        math_data = []
+        mathml_data, latex_data = ocr.math_data_parts(math_data)
+        self.assertEqual(mathml_data, None)
+        self.assertEqual(latex_data, None)
